+++ include/libparasheet/foo.h
@@ -1,0 +1,1 @@
+void foo(void);
