+++ src/libparasheet/foo.c
@@ -1,0 +1,6 @@
+#include <stdio.h>
+#include <libparasheet/foo.h>
+
+void foo(void) {
+  puts("foo");
+}
