+++ src/parasheet-cli/main.c
@@ -1,0 +1,3 @@
+#include <libparasheet/foo.h>
+
+int main() { foo(); }
