+++ src/parasheet-editor/main.c
@@ -1,4 +1,4 @@
-#include <util.h>
+#include <util/util.h>
 
 int main() {
 
@@ -40,3 +40,4 @@ int main() {
 
     StackAllocatorDestroy(&a);
 }
+
